# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZenKart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user's orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail and tracking
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # WISHLIST
...
  /listings:
    get:
      operationId: listUsedGoods
      tags: [listings]
      summary: List C2C used goods
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: condition
          in: query
          schema:
            type: string
            enum: [new, like_new, good, fair]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: city
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a used goods listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{listingId}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing detail
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /listings/{listingId}/make-offer:
    post:
      operationId: makeOffer
      tags: [listings]
      summary: Make an offer on a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MakeOfferBody"
      responses:
        "200":
          description: Offer sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # RESELLER
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productTitle:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
        variant:
          type: string
          nullable: true
      required: [productId, productTitle, productImage, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        deliveryAddress:
          type: string
        expectedDelivery:
          type: string
        createdAt:
          type: string
      required: [id, userId, items, subtotal, total, status, paymentMethod, paymentStatus, deliveryAddress, createdAt]
    TrackingEvent:
      type: object
      properties:
        status:
          type: string
        description:
          type: string
        timestamp:
          type: string
        isCompleted:
          type: boolean
      required: [status, description, timestamp, isCompleted]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            trackingEvents:
              type: array
              items:
                $ref: "#/components/schemas/TrackingEvent"
            trackingNumber:
              type: string
              nullable: true
            courierName:
              type: string
              nullable: true
          required: [trackingEvents]
    CreateOrderBody:
      type: object
      properties:
        addressId:
          type: integer
        paymentMethod:
          type: string
          enum: [upi, card, netbanking, cod, wallet, emi, bnpl]
        couponCode:
          type: string
      required: [addressId, paymentMethod]
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        sellerName:
          type: string
        sellerAvatar:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        price:
          type: number
        isNegotiable:
          type: boolean
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        condition:
          type: string
          enum: [new, like_new, good, fair]
        city:
          type: string
        state:
          type: string
        viewCount:
          type: integer
        isActive:
          type: boolean
        isFree:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, sellerName, title, description, price, isNegotiable, images, categoryId, condition, city, state, viewCount, isActive, isFree, createdAt]
    ListingListResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [listings, total, page, totalPages]
    CreateListingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        isNegotiable:
          type: boolean
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        condition:
          type: string
          enum: [new, like_new, good, fair]
        city:
          type: string
        state:
          type: string
        isFree:
          type: boolean
      required: [title, description, price, categoryId, condition, city, state]
    MakeOfferBody:
      type: object
      properties:
        offerPrice:
          type: number
        message:
          type: string
      required: [offerPrice]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [order, deal, system, price_drop, message]
        isRead:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        actionUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, title, body, type, isRead, createdAt]
...